reflect: add Kind, remove Int8Type, Int8Value, etc.
update other code to match.

R=r
CC=golang-dev
https://golang.org/cl/1680044

+++ src/pkg/exp/ogle/rruntime.go
@@ -265,7 +265,7 @@ func fillRuntimeIndexes(runtime *runtimeValues, out *runtimeIndexes) {
 		for j := 0; j < outStructt.NumField(); j++ {
 			f := outStructv.Field(j).(*reflect.IntValue)
 			name := outStructt.Field(j).Name
-			f.Set(indexes[name])
+			f.Set(int64(indexes[name]))
 		}
 	}
 }
